Revert "FIx type"
This reverts commit 7d8e17570fcb377eb51a23b74d734231e45cffd1.

+++ include/seqan/kmer/kmer_ibf.h
@@ -224,7 +224,7 @@ public:
      */
     void select(std::vector<uint16_t> & counts, TString const & text) // TODO uint16_t
     {
-        uint32_t possible = length(text) - kmerSize + 1; // Supports text lengths up to 65535 + k
+        uint16_t possible = length(text) - kmerSize + 1; // Supports text lengths up to 65535 + k
 
         std::vector<uint64_t> kmerHashes(possible, 0);
 
@@ -232,7 +232,7 @@ public:
         resize(kmerShape, kmerSize);
         hashInit(kmerShape, begin(text));
         auto it = begin(text);
-        for (uint32_t i = 0; i < possible; ++i)
+        for (uint16_t i = 0; i < possible; ++i)
         {
             kmerHashes[i] = hashNext(kmerShape, it);
             ++it;
